requirements.txt
python-dotenv
requests
pydantic
replicate
openai
kokoro>=0.3.1
soundfile
langgraph
langchain-groq
moviepy==1.0.3